+++ frontend/src/components/eventPage/EventCard.jsx
